+++ server/controllers/viajes.controllers.js
@@ -21,6 +21,7 @@ export const getViajes = async (req, res) => {
         res.json(result)
     } catch (error) {
         console.log(error)
+        res.status(500).json({ message: "Error al obtener los viajes" })
     }
 }
 
@@ -31,12 +32,18 @@ export const getViaje = async (req, res) => {
         res.json(result[0])
     } catch (error) {
         console.log(error)
+        res.status(500).json({ message: "Error al obtener los viajes" })
     }
 }
 
 export const createViaje = async (req, res) => {
     try {
         const {idViaje, Origen, Destino, idCamion, fecha, hora,precio } = req.body
+        const [existingViajes] = await pool.query('SELECT * FROM viaje WHERE idViaje = ? OR idCamion = ?', [idViaje, idCamion]);
+        if (existingViajes.length > 0) {
+            return res.status(400).json({ message: "Ya existe una viaje con el mismo id o camion asignado"})
+        }
+
         const [origenRes] = await pool.query('SELECT idCiudad FROM ciudad where estado = ?', [Origen])
         const idOrigen = origenRes[0].idCiudad
         const [destinoRes] = await pool.query('SELECT idCiudad FROM ciudad where estado = ?', [Destino])
@@ -53,12 +60,19 @@ export const createViaje = async (req, res) => {
         })
     } catch (error) {
         console.log(error)
+        res.status(500).json({ message: "Error al registrar el viaje" })
     }
 }
 
 export const updateViaje = async (req, res) => {
     try {
         const {Origen,Destino,idCamion,fecha,hora} = req.body
+        const [existingViajes] = await pool.query('SELECT * FROM viaje WHERE idViaje = ? OR idCamion = ?', [idViaje, idCamion]);
+        if (existingViajes.length > 0) {
+            return res.status(400).json({ message: "Ya existe una viaje con el mismo id o camion asignado"})
+        }
+
+
         const [origenRes] = await pool.query('SELECT idCiudad FROM ciudad where estado = ?', [Origen])
         const idOrigen = origenRes[0].idCiudad
         const [destinoRes] = await pool.query('SELECT idCiudad FROM ciudad where estado = ?', [Destino])
@@ -68,13 +82,19 @@ export const updateViaje = async (req, res) => {
         console.log(result)
     } catch (error) {
         console.log(error)
+        res.status(500).json({ message: "Error al actualizar el viaje" })
     }
 }
 
 export const deleteViaje = async (req, res) => {
     try {
+        const [viajeResult] = await pool.query('SELECT * FROM boleto WHERE idViaje = ?', [req.params.id])
+        if (viajeResult.length > 0) {
+            return res.status(400).json({ message: "No se puede eliminar el viaje porque está asociado con uno o más boletos." })
+        }
+
         const [result] = await pool.query('DELETE FROM viaje WHERE idViaje = ?', [req.params.id])
-        if (result.affectedRows === 0) return res.status(404).json({ message: "ruta not found" })
+        if (result.affectedRows === 0) return res.status(404).json({ message: "viaje no encontrado" })
         return res.sendStatus(204)
     } catch (error) {
         console.log(error)
